+++ FilmSearch/app/src/main/java/com/example/filmsearch/viewmodel/HomeFragmentViewModel.kt
@@ -5,12 +5,14 @@ import androidx.lifecycle.ViewModel
 import com.example.filmsearch.App
 import com.example.filmsearch.domain.Film
 import com.example.filmsearch.domain.Interactor
+import okhttp3.internal.Internal.instance
+import org.koin.core.KoinComponent
+import org.koin.core.inject
 
-class HomeFragmentViewModel : ViewModel() {
-    val filmsListLiveData: MutableLiveData<List<Film>> = MutableLiveData()
-
+class HomeFragmentViewModel : ViewModel(), KoinComponent {
+    val filmsListLiveData:  MutableLiveData<List<Film>> = MutableLiveData()
     //Инициализируем интерактор
-    private var interactor: Interactor = App.instance.interactor
+    private val interactor: Interactor by inject()
 
     init {
         interactor.getFilmsFromApi(1, object : ApiCallback {
