+++ FilmSearch/readme.md
@@ -4,8 +4,6 @@
 
 ### Задание:
 
-Сделать загрузку фильмов с помощью _Retrofit_ в _RecyclerView_ на главном экране приложения.
+Перевести проект на Koin.
 
-✱ Реализовать пагинацию. (Пагинация в этом случае значит, что когда пользователь доходит до конца загруженной страницы, должна быть загружена следующая страница. Желательно загрузка следующей страницы должна начаться незадолго до конца предыдущей, чтобы создать эффект «бесшовности».)
-
-![Скриншот Модуль 35](./pix/FilmSearch_35.png)
+✱Перевести проект на Hilt.
